Increase member quorum requirement from 5% to 20%
Given the size of the coop, the board felt that no substantial changes
to the Cooperative should be possible with only 5% of members voting
affirmatively. At the same time, the board is cognizant of the fact that
low meeting attendance combined with a high quorum requirement can make
it difficult for the Cooperative to conduct its business.

Members approved a board proposal to raise the quorum requirement during
the 2013 Annual Meeting. Per board discussion, and taking into account
member input on the issue, raising the quorum requirement to 20% was
decided as a good balance.

As approved by the board on 2013-09-18.